refactor: page token implementation

+++ src/Lab2.Tools.Abstractions/Persistence/ConfigurationQuery.cs
@@ -3,4 +3,4 @@ using SourceKit.Generators.Builder.Annotations;
 namespace Lab2.Tools.Abstractions.Persistence;
 
 [GenerateBuilder]
-public partial record ConfigurationQuery([RequiredValue] int PageSize, [RequiredValue] int Cursor);
+public partial record ConfigurationQuery([RequiredValue] int PageSize, [RequiredValue] long Cursor);
